+++ demo/admin/src/products/future/ProductForm.cometGen.ts
@@ -25,6 +25,16 @@ export const ProductForm: FormConfig<GQLProduct> = {
                 { type: "text", name: "description", label: "Description", multiline: true },
                 { type: "staticSelect", name: "type", label: "Type", required: true, values: [{ value: "Cap", label: "great Cap" }, "Shirt", "Tie"] },
                 { type: "asyncSelect", name: "category", rootQuery: "productCategories" },
+                {
+                    type: "optionalNestedFields",
+                    name: "dimensions",
+                    checkboxLabel: "Configure dimensions",
+                    fields: [
+                        { type: "number", name: "width", label: "Width" },
+                        { type: "number", name: "height", label: "Height" },
+                        { type: "number", name: "depth", label: "Depth" },
+                    ],
+                },
             ],
         },
         {
@@ -34,6 +44,8 @@ export const ProductForm: FormConfig<GQLProduct> = {
                 { type: "boolean", name: "inStock" },
                 { type: "date", name: "availableSince" },
                 { type: "block", name: "image", label: "Image", block: { name: "DamImageBlock", import: "@comet/cms-admin" } },
+                { type: "fileUpload", name: "priceList", label: "Price List", maxFileSize: 1024 * 1024 * 4 },
+                { type: "fileUpload", name: "datasheets", label: "Datasheets", multiple: true, maxFileSize: 1024 * 1024 * 4 },
             ],
         },
     ],