+++ packages/admin/admin/src/EditDialog.test.tsx
@@ -1,6 +1,5 @@
 import { Button } from "@mui/material";
 import userEvent from "@testing-library/user-event";
-import * as React from "react";
 import { render, screen, waitFor } from "test-utils";
 
 import { useEditDialog } from "./EditDialog";