+++ packages/admin/admin/src/dataGrid/useBufferedRowCount.tsx
@@ -1,11 +1,11 @@
-import * as React from "react";
+import { useEffect, useState } from "react";
 
 export function useBufferedRowCount(rowCount: number | undefined) {
     // Some API clients return undefined while loading
     // Following lines are here to prevent `rowCountState` from being undefined during the loading
-    const [rowCountState, setRowCountState] = React.useState(0);
+    const [rowCountState, setRowCountState] = useState(0);
 
-    React.useEffect(() => {
+    useEffect(() => {
         setRowCountState((prevRowCountState) => (rowCount !== undefined ? rowCount : prevRowCountState));
     }, [rowCount, setRowCountState]);