+++ packages/admin/admin/src/router/Prompt.tsx
@@ -1,5 +1,5 @@
 import * as History from "history";
-import * as React from "react";
+import { createContext, PropsWithChildren, useContext, useEffect, useRef } from "react";
 import { __RouterContext } from "react-router";
 import useConstant from "use-constant";
 import { v4 as uuid } from "uuid";
@@ -16,7 +16,7 @@ interface PromptContext {
     register: (id: string, path: string) => void;
     unregister: (id: string) => void;
 }
-export const PromptContext = React.createContext<PromptContext | undefined>(undefined);
+export const PromptContext = createContext<PromptContext | undefined>(undefined);
 
 // react-router Prompt doesn't support multiple Prompts, this one does
 interface IProps {
@@ -29,17 +29,17 @@ interface IProps {
     resetAction?: ResetAction;
     subRoutePath?: string;
 }
-export const RouterPrompt: React.FunctionComponent<IProps> = ({ message, saveAction, resetAction, subRoutePath, children }) => {
+export const RouterPrompt = ({ message, saveAction, resetAction, subRoutePath, children }: PropsWithChildren<IProps>) => {
     const id = useConstant<string>(() => uuid());
-    const reactRouterContext = React.useContext(__RouterContext); // reactRouterContext can be undefined if no router is used, don't fail in that case
+    const reactRouterContext = useContext(__RouterContext); // reactRouterContext can be undefined if no router is used, don't fail in that case
     const path: string | undefined = reactRouterContext?.match?.path;
-    const context = React.useContext(RouterContext);
+    const context = useContext(RouterContext);
     const subRoutePrefix = useSubRoutePrefix();
-    const promptRoutes = React.useRef<PromptRoutes>({});
+    const promptRoutes = useRef<PromptRoutes>({});
     if (subRoutePath && subRoutePath.startsWith("./")) {
         subRoutePath = subRoutePrefix + subRoutePath.substring(1);
     }
-    React.useEffect(() => {
+    useEffect(() => {
         if (context) {
             context.register({ id, message, saveAction, resetAction, path, subRoutePath, promptRoutes });
         } else {